requirements.txt
flask==3.0.2
openai-whisper==20231117
numpy==1.26.4
torch==2.2.1
ffmpeg-python==0.2.0
gunicorn==21.2.0 